Add files via upload
"We are happy to present the result of our labors: a machine learning system that has the potential to completely transform the prediction of cardiac disease. With painstaking data analysis and model optimization, we have developed a technology that provides unmatched cardiovascular risk assessment accuracy. Come usher in a new era of healthcare that is tailored to each individual. Predictive Heart Health, AI in Healthcare, and Precision Medicine.